% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classiccompare-package.R
\docType{package}
\name{classiccompare-package}
\alias{classiccompare}
\alias{classiccompare-package}
\title{classiccompare: Compare Two Datasets and Report Differences}
\description{
Compares a base and a comparison data frame and reports differences in variables, observations and values, in the style of the SAS dataset comparison utility. Observations can be matched by position, ID or BY variables. Values are compared exactly or with absolute, relative or percent criteria, and a bitmask return code supports automated quality control checks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sharangnimonkar/clinical-programming-portfolio/tree/main/01_R_Package_Development}
  \item Report bugs at \url{https://github.com/sharangnimonkar/clinical-programming-portfolio/issues}
}

}
\author{
\strong{Maintainer}: Sharang Nimonkar \email{sharangnimonkar@gmail.com}

}
\keyword{internal}
