% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysinfo.R
\name{sysinfo_decode}
\alias{sysinfo_decode}
\title{Names of the conditions set in a return code}
\usage{
sysinfo_decode(code)
}
\arguments{
\item{code}{An integer return code from \code{\link[=sysinfo]{sysinfo()}}.}
}
\value{
A character vector of condition names, empty for 0.
}
\description{
Names of the conditions set in a return code
}
\examples{
sysinfo_decode(4096L + 64L)
}
